backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
opencv-python==4.8.1.78
numpy==1.26.2
Pillow==10.0.0
openai==1.3.5
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
pydantic==2.5.2
pytest==7.4.3
httpx==0.25.2 